requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Web Scraping
playwright==1.41.0
playwright-stealth==1.0.6
beautifulsoup4==4.12.3
lxml==5.1.0

# AI / LLM
google-generativeai==0.3.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Scheduling
apscheduler==3.10.4

# Data Processing
pydantic==2.5.3
python-dateutil==2.8.2

# Utilities
aiofiles==23.2.1
jinja2==3.1.3
